Move command traits to feature namespace
