+++ module/VuFindSearch/src/VuFindSearch/Command/RetrieveCommand.php
@@ -31,6 +31,7 @@
 namespace VuFindSearch\Command;
 
 use VuFindSearch\Backend\BackendInterface;
+use VuFindSearch\Command\Feature\RecordIdentifierTrait;
 use VuFindSearch\ParamBag;
 
 /**
